Rename files and correct namespaces from OctopusTools -> Octopus.Cli
