+++ source/Octopus.Cli/Commands/IChannelResolver.cs
@@ -6,7 +6,7 @@ using System.Threading.Tasks;
 using Octopus.Client;
 using Octopus.Client.Model;
 
-namespace OctopusTools.Commands
+namespace Octopus.Cli.Commands
 {
     public interface IChannelResolver
     {
