+++ source/Octopus.Cli/Octopus.Cli.csproj
@@ -103,6 +103,7 @@
     <Compile Include="..\Solution Items\VersionInfo.cs">
       <Link>Properties\VersionInfo.cs</Link>
     </Compile>
+    <Compile Include="Commands\ChannelResolverHelper.cs" />
     <Compile Include="Commands\CleanEnvironmentCommand.cs" />
     <Compile Include="Commands\CreateEnvironmentCommand.cs" />
     <Compile Include="Commands\IPackageBuilder.cs" />
